exp/ingest: Polish interfaces
