+++ exp/ingest/adapters/mocks.go
@@ -2,6 +2,7 @@ package adapters
 
 import (
 	"context"
+
 	"github.com/stellar/go/exp/ingest/io"
 	"github.com/stellar/go/xdr"
 	"github.com/stretchr/testify/mock"
@@ -22,13 +23,8 @@ func (m *MockHistoryArchiveAdapter) BucketListHash(sequence uint32) (xdr.Hash, e
 }
 
 func (m *MockHistoryArchiveAdapter) GetState(
-	ctx context.Context, sequence uint32, tempSet io.TempSet, maxStreamRetries int,
-) (io.StateReader, error) {
-	args := m.Called(ctx, sequence, tempSet, maxStreamRetries)
-	return args.Get(0).(io.StateReader), args.Error(1)
-}
-
-func (m *MockHistoryArchiveAdapter) GetLedger(sequence uint32) (io.ArchiveLedgerReader, error) {
-	args := m.Called(sequence)
-	return args.Get(0).(io.ArchiveLedgerReader), args.Error(1)
+	ctx context.Context, sequence uint32, maxStreamRetries int,
+) (io.ChangeReader, error) {
+	args := m.Called(ctx, sequence, maxStreamRetries)
+	return args.Get(0).(io.ChangeReader), args.Error(1)
 }
