+++ exp/ingest/io/errors.go
@@ -1,0 +1,7 @@
+package io
+
+import (
+	"github.com/stellar/go/support/errors"
+)
+
+var ErrNotFound = errors.New("ledger not found")
