+++ exp/ingest/io/mock_change_reader.go
@@ -14,3 +14,8 @@ func (m *MockChangeReader) Read() (Change, error) {
 	args := m.Called()
 	return args.Get(0).(Change), args.Error(1)
 }
+
+func (m *MockChangeReader) Close() error {
+	args := m.Called()
+	return args.Error(0)
+}
